videobackend/requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
avro==1.11.3
billiard==4.2.0
boto3==1.34.34
botocore==1.34.34
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
decorator==4.4.2
dotenv-python==0.0.1
fastapi==0.109.2
h11==0.14.0
humps==0.2.2
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
jmespath==1.0.1
kafka-python==2.0.2
kombu==5.3.5
moviepy==1.0.3
numpy==1.26.3
Pillow==9.5.0
proglog==0.1.10
prompt-toolkit==3.0.43
pydantic==2.6.0
pydantic_core==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
s3transfer==0.10.0
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
uvicorn==0.27.0.post1
vine==5.1.0
wcwidth==0.2.13
